Dump the entire process tree when calling KillTree
When we time out/give up/abort we should print out
the entire process tree to help us figure out which
processes are still running.

This will help diagnose issues where one process is
waiting on a 'Process.WaitForExit' and we're not sure
if the child process has actually exited or not.