Readd old API endpoints in WebServices, reload lane info from DB in WebServices to check permissions
I made the mistake of assuming that WebMethod was not an API endpoint method. My changes also _broke_ the API. This should restore it back to what it was and for the new methods we added actually check the permissions of the lane in the method itself, so no one should be able to screw with it.