Change GitHub Oauth to use Teams rather than Organizations
Originally I set this up to just use GitHub Organizations for auth. The idea was that if you were that organization, you would get whatever roles were required. However that logic fails because users could be apart of an org who we don't want to have full access. So we need a more fine control. Enter: Teams! Now you need to be apart of the organization AND a team in order to get full access.
In GitHubOrganizationList in the MonkeyWrench Config, the schema is `<org>*<team,seperated by commas>:<Roles>`