Refactor GitHub login to make it simpler
The code for logging in with GitHub, while functional, was a bit hard to understand. So to make it easier and cleaner, I refactored it. The old login methods for OpenID/Google are back to how they were, while GitHub has its own method for logging in, with much simpler logic for handling it. It also has its own error message should the user not be able to log in.