Fix extra whitespace in AssemblyVersion
When generating the AssemblyInfo.cs file we end up with extra whitespace:

[assembly: AssemblyVersion ("1.0.0.     787")]

This compiled fine in previous versions of mono. But in 4.9.0 this is
enforced as invalid.

AssemblyInfo.cs(27,29): error CS7034: The specified version string does
not conform to the required format - major[.minor[.build[.revision]]]

This removes the extra whitespace and makes the version number valid.